Efforts to get the form layout layout correctly.
Now passing in a label width and I've hardcoded the field width as I
can't find a way to make the field width to expand to fill the space and
the vaadin document tends to support this being an issue.

I'm thinking we need the ability to set label and field widths on a per
column basis to give is the desired level of control.